+++ hdd/2nd_week/team_collaboration_tool/bin/data/data_sources/team_dao.dart
@@ -1,24 +1,18 @@
-import 'package:drift/drift.dart';
+import 'package:riverpod/riverpod.dart';
 
 import '../../db/database.dart';
-import '../entities/team_table.dart';
+import 'team_dao_impl.dart';
 
-part 'team_dao.g.dart';
+final teamDaoProvider = Provider<TeamDao>(
+  (ref) => TeamDaoImpl(
+    ref.watch(databaseProvider),
+  ),
+);
 
-@DriftAccessor(tables: [TeamTable])
-class TeamDao extends DatabaseAccessor<Database> with _$TeamDaoMixin {
-  TeamDao(Database db) : super(db);
+abstract class TeamDao {
+  Future<TeamEntity?> getTeamOrNull(String name);
 
-  Future<TeamEntity?> getTeamOrNull(String name) {
-    return (db.teamTable.select()..where((team) => team.name.equals(name)))
-        .getSingleOrNull();
-  }
+  Future<int> insertTeam(String name);
 
-  Future<int> insertTeam(String name) {
-    return into(teamTable).insert(TeamEntity(name: name));
-  }
-
-  Future<List<TeamEntity>> getTeams() {
-    return db.teamTable.select().get();
-  }
+  Future<List<TeamEntity>> getTeams();
 }
