+++ hdd/2nd_week/team_collaboration_tool/bin/data/data_sources/team_dao_impl.dart
@@ -1,0 +1,29 @@
+import 'package:drift/drift.dart';
+
+import '../../db/database.dart';
+import '../entities/team_table.dart';
+import 'team_dao.dart';
+
+part 'team_dao_impl.g.dart';
+
+@DriftAccessor(tables: [TeamTable])
+class TeamDaoImpl extends DatabaseAccessor<Database>
+    with TeamDao, _$TeamDaoImplMixin {
+  TeamDaoImpl(Database db) : super(db);
+
+  @override
+  Future<TeamEntity?> getTeamOrNull(String name) {
+    return (db.teamTable.select()..where((team) => team.name.equals(name)))
+        .getSingleOrNull();
+  }
+
+  @override
+  Future<int> insertTeam(String name) {
+    return into(teamTable).insert(TeamEntity(name: name));
+  }
+
+  @override
+  Future<List<TeamEntity>> getTeams() {
+    return db.teamTable.select().get();
+  }
+}
