+++ hdd/2nd_week/team_collaboration_tool/bin/data/repositories/team_repository_impl.dart
@@ -1,0 +1,29 @@
+import '../../db/database.dart';
+import '../data_sources/team_dao.dart';
+import 'team_repository.dart';
+
+class TeamRepositoryImpl implements TeamRepository {
+  final TeamDao teamDao;
+
+  TeamRepositoryImpl(this.teamDao);
+
+  @override
+  deleteTeam() {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<TeamEntity?> getTeamOrNull(String teamName) {
+    return teamDao.getTeamOrNull(teamName);
+  }
+
+  @override
+  Future<int> insertTeam(String teamName) {
+    return teamDao.insertTeam(teamName);
+  }
+
+  @override
+  Future<List<TeamEntity>> getTeams() {
+    return teamDao.getTeams();
+  }
+}
