+++ hdd/2nd_week/team_collaboration_tool/bin/db/database.dart
@@ -1,11 +1,15 @@
 import 'package:drift/drift.dart';
+import 'package:drift/native.dart';
+import 'package:riverpod/riverpod.dart';
 import 'package:uuid/uuid.dart';
-import '../data/data_sources/team_dao.dart';
+import '../data/data_sources/team_dao_impl.dart';
 import '../data/entities/team_table.dart';
+
 part 'database.g.dart';
 
+final databaseProvider = Provider<Database>((ref) => Database(NativeDatabase.memory()));
 
-@DriftDatabase(tables: [TeamTable], daos: [TeamDao])
+@DriftDatabase(tables: [TeamTable], daos: [TeamDaoImpl])
 class Database extends _$Database {
   Database(QueryExecutor e) : super(e);
 
@@ -14,8 +18,6 @@ class Database extends _$Database {
 
   @override
   MigrationStrategy get migration {
-    return MigrationStrategy(
-
-    );
+    return MigrationStrategy();
   }
 }
