+++ hdd/2nd_week/team_collaboration_tool/bin/domain/services/team_service.dart
@@ -1,5 +1,14 @@
+import 'package:riverpod/riverpod.dart';
+
 import '../../data/repositories/team_repository.dart';
 import '../models/team.dart';
+import 'team_service_impl.dart';
+
+final teamServiceProvider = Provider<TeamService>(
+  (ref) => TeamServiceImpl(
+    ref.watch(teamRepositoryProvider),
+  ),
+);
 
 abstract class TeamService {
   Future<int> addTeam(String teamName);
@@ -9,37 +18,4 @@ abstract class TeamService {
   dynamic removeTeam();
 
   Future<List<Team>> getTeamList();
-}
-
-class TeamServiceImpl implements TeamService {
-  final TeamRepository _teamRepository;
-
-  TeamServiceImpl(this._teamRepository);
-
-  @override
-  Future<int> addTeam(String teamName) async {
-    return await _teamRepository.insertTeam(teamName);
-  }
-
-  @override
-  Future<Team?> searchTeamByName(String teamName) async {
-    var result = await _teamRepository.getTeamOrNull(teamName);
-
-    if (result != null) {
-      return Team(id: result.id!, name: result.name);
-    }
-    return null;
-  }
-
-  @override
-  Future<List<Team>> getTeamList() async {
-    var result = await _teamRepository.getTeams();
-    return result.map((e) => Team.fromEntity(e)).toList();
-  }
-
-  @override
-  dynamic removeTeam() {
-    // TODO: implement removeTeam
-    throw UnimplementedError();
-  }
 }
